+++ src/components/common/Form/input/index.tsx
@@ -3,7 +3,7 @@
 import { clsx } from "clsx";
 import { ComponentProps, forwardRef, useState } from "react";
 
-import useToggle from "@/hooks/useToggle";
+import useToggle from "@/hooks/use-toggle";
 import { IconVisibilityOff, IconVisibilityOn } from "@/public/assets/icons";
 
 interface InputProps extends ComponentProps<"input"> {
