+++ src/components/common/drop-down/index.stories.tsx
@@ -2,7 +2,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { type ComponentType } from "react";
 
-import useToggle from "@/hooks/useToggle";
+import useToggle from "@/hooks/use-toggle";
 
 import DropDown from ".";
